+++ src/main/java/cn/aprilviolet/highlightbracketpair/setting/HighlightBracketPairPluginState.java
@@ -1,0 +1,24 @@
+package cn.aprilviolet.highlightbracketpair.setting;
+
+/**
+ * Idea中配置颜色设置的状态类
+ *
+ * @author AprilViolet
+ * @version V1.0.0
+ * @date 2021.07.31 星期六 11:02
+ * @since V1.0.0
+ */
+public class HighlightBracketPairPluginState {
+    /**
+     * 插件版本
+     */
+    private String pluginVersion = "";
+
+    public String getPluginVersion() {
+        return pluginVersion;
+    }
+
+    public void setPluginVersion(String pluginVersion) {
+        this.pluginVersion = pluginVersion;
+    }
+}